Added throwing of TransientException to authorizer, metadatasource interfaces.
git-svn-id: https://opencadc.googlecode.com/svn/trunk@1755 728ff76a-78ac-11de-a72b-d90af8dea425

+++ projects/cadcUtil/src/ca/nrc/cadc/net/TransientException.java
@@ -118,7 +118,7 @@ public class TransientException extends Exception
      * Get the retry delay (in seconds)
      * @return
      */
-    public long getRetryDelay()
+    public int getRetryDelay()
     {
         return retryDelay;
     }
